Major cleanup, optimization across the board and improvements to get ready for the next release.

+++ arm9/source/emu/Pokey.c
@@ -71,7 +71,7 @@
 
 byte pokey_buffer[SNDLENGTH] = {0};
 
-u16 pokeyBufIdx __attribute__((section(".dtcm"))) = 0;
+u32 pokeyBufIdx __attribute__((section(".dtcm"))) = 0;
 
 static uint pokey_frequency = 1787520;
 static uint pokey_sampleRate = (31440/2);
